Fix openMultipleChannels function
The openMultipleChannels function had an unnoticed bug that was related to comparing enough balance to open the desired channels. Using the ethers package we can ensure that the numbers representing the wei values can be accurately multiplied and compared. The wei values can be multiplied and compared accurately.

In addition to this change, we have added a Promise.allSettled to make sure that all channels are open before returning a result.